refactor: move box plot client
